Attempted text size tweak for mobile portrait orientation
This won't salvage our floated content, and we can't reposition that content very well because floats bad, but hopefully we can at least make text readable at the default zoom.

There's no way to directly detect mobile browsers. Media queries and "responsive design" weren't designed for and aren't adequate for the high-resolution, small-size smartphones of today. This media query is basically a blind guess, and even if it works today, it'll probably break a few years down the line.