Troubleshooting floated content width
Attempting to set a sensible maximum width on floated content so that on mobile devices in portrait orientation, we don't end up with situations where all text is wrapped below the floats even as the floats only take up 60% or 70% of the containing card's width.

Added clearfix to the section cards. Just in case.