Improvements for smartphones; improved favicon; 404 page
When viewed on a smartphone in portrait orientation, we un-float the asides, pulling them below the first paragraph of the containing card. In order to allow all of them -- including the YouTube embeds -- to properly fill the card's width, we have to use inline styles on the embeds to set their CSS `aspect-ratio`.

Modified the main layout file to allow pages to specify additional CSS files to include.

Added a 404 page with a dusk background theme.

Made the outlines in the favicon thicker, and softened the inner line and the moon's shadow.

Removed an old, unused layout file.

Documented the CSS media query used to target smartphones.