Extensive work, to add articles to the site
This task ballooned into a lot more work than I expected, and it didn't occur to me to branch it until now, so this commit will be a bit... disgustingly large.

I'm currently trying to add an articles section to the site, as a secondary focus to the portfolio (so, the only link from the homepage is in the site footer). Initial articles will include writings on raycast math, to eventually include public-domain code samples.

I intend to publish these changes in full once I have at least two articles: a guide for raycasting cones, and a guide for raycasting cylinders. (Still need to work out the math for the latter. I've been using the articles themselves as notepads of sorts.) I might also take my math cheatsheet (containing things like FOIL) and make a third article out of it.

---

Site HEAD content (social media META tags, favicon, etc.) moved into a separate Liquid include, to allow reuse across multiple layouts.

Separated out my main SCSS file into utilities, theming constants, and layout information.

Refactored the CSS background, which previously consisted of several gradients (to form the sand-and-sky coloration) and a pseudo-element (for scanlines). All backgrounds are now defined directly on the HTML element as layers. Some of the new SCSS utilities are used to properly merge background layers, so that the scanlines and gradients can be coded separately. (This, in turn, allows me to override just the gradients, so the 404 page now does so.)

Created a layout with a sidebar, for use on articles and the article listing page. This layout rearranges itself on smartphones (with some ugly accommodations to work around a Firefox bug with CSS media queries). The layout isn't perfect (among other things, the "site nav" portion of the sidebar is always default-collapsed), but it is what it needs to be and does what it needs to do in order to work optimally across platforms without any JavaScript.

Created an (ugly) system for mathematical markup, to allow for math in HTML, without JavaScript, that copies properly to plaintext and is only slightly a disaster for screen readers. There is ample room for improvement in this regard, but also ample possibility to improve. The system is called `lu-math`. Full documentation pending.

Added a site footer which links to the articles page. In the sidebar layout, on mobile, social links are moved down here, too.

Added a draft of an article on computing ray/cone intersections. The article is incomplete: I still need to test my C++ implementation (by porting it to GLSL and throwing it into Shadertoy for a raytracing test) and, if it works, include it in the article under CC0/public domain.