CSS rowSpan article; started ray/cylinder article; site theming
Wrote an article on CSS's inability to cope with rowSpan, after running into it while trying to theme the site. I wonder if it's the kind of thing that could be sent to the CSSWG.

Wrote site styles for tables, and for "multi-figures."

Pages in a collection (i.e. articles) can now include stylesheets from a subfolder named after the page.

The lu-math stuff now has functionality for binomial coefficients and for summation.

Code tags in a page's "grand header" now have some extra styling, so that the change in typography hopefully won't look as weird.

Started work on deriving ray/cylinder intersection code.