Table theming improvements
Rowspan fixes to tables in the article on the multinomial theorem.

Improved table styles: fixed edge-case problems for table corners, and added more features and functionality -- giving us a thick black-outlined dark-grey border on tables now.

The border-collapse mixin now accounts for the fact that TRs are always wrapped in TBODYs if placed directly in a TABLE, i.e. a TR will never be a TABLE's direct child.