Update css-rowspan-issue.html
Updated the intro blurb on this article.

This article went through a lot of revisions, as I dealt with issues with styling tables. There are weaknesses in the CSS spec that obscure deeper issues; there are layers of jank you have to dig through to get at the heart of some of these problems; so the article evolved as I wrote, until eventually, it was about a completely different problem than it started off at.