math
Work on the ray/cylinder derivation: I restarted from an earlier point in the process.

Work on lu-math:

Created a lu-math-multiline element that can be used to break a long equation into multiple lines. The equation will be displayed as three columns, with top content in the left column, multiple lines in the middle column, and bottom content in the right column. Word-wrapping in the left column properly pushes the middle column down. Unfortunately, due to limitations in CSS and how it exposes line-box-sizing primitives, this requires knowledge of the font metrics.

The "enormous rows" container now uses dotted lines instead of solid ones.

Fractions now use grids instead of flexbox, and no longer use a pseudo-element for the divider line.

A lu-math-big-parens that is the direct child of a MARK will now inherit the MARK's background color, as a cheap hack to get basic highlighting of these elements working. Square roots and other multi-line containers would also need this fix, but I've not yet applied it. In general, I want to redesign lu-math at some point; so far, it's evolved (too) organically rather than been built with a real design in mind.