lu-math; ray/cylinder
Ray/cylinder:

Restarted the derivation from my last known mistake. Made it to the end, but the equation is still incredibly unwieldy.

lu-math2:

Changed the layout for the "snake" layout to be a little less janky. Sadly, it won't be possible to make this actually be good until browsers implement CSS baseline-source, which was added to the CSS Inline 3 draft in 2019 but isn't even on Can I Use or MDN.

Separated out the behavior that was built into the "snake" layout's body section: this behavior is now a new layout mode, "v-segments."

Text in a solvestep's explanation is now left-aligned.

Marks in math are now only announced to screen readers if a class is set on a containing element. If highlighting is referred to in text, then this class should be used.