More math, as well as math theming
Ray/cylinder derivation attempts: Giving this a try using a cross product and identities.

---

lu-math2:

Updated styles for exponentiation.

Created a more general set of styles for elements that represent groupings with vertically-resizing delimiters. This is now used for vector length operations, and for `m-big-parens` elements. None of my implementations for this have ever been able to cope with word-wrapping, so I'm cutting my losses there and making these groupings use CSS grids. This allows inline elements with backgrounds and borders, like MARK and DEL, to properly wrap them.

Made it possible to decide whether an explanation expander button is locked to the bottom of an equation, or centered along the equation's height.

Fixed vertical text alignments for givens lists, when a definition word wraps.

Updated styles for fractions. These also use grids, and now have code to properly handle raw and rawcap elements inside the fraction.

MARK and SPAN elements inside of lu-math are now inline-block by default.

Fixed vertical sizing and margins on the block-level "therefore" element.

---

Site theming:

Expand/collapse buttons for solvestep explanations are now displayed as a circular badge with a question mark on it.

When a solvestep explanation is expanded, any deletions in the relevant solvestep (indicated with DEL elements) are now made visible, with a striped red background.

Text alignment within solvestep explanations is now consistently left, unjustified.