Work on ray/triangle math; site math theming tweaks
Additions to lu-math to allow for forced-lowercase on terms (so that screen readers (hopefully) read them as "AB" i.e. one letter at a time, but visually they are lowercase), and to allow for large delimiters (i.e. large commas within large parentheses).

Fixes to the article theming to allow for multiple SVG-based graphs in a single figure.

Started work on an article for ray/triangle math. GLM's function doesn't allow you to tell whether you're hitting the triangle's backface, so it doesn't meet my needs.