Update targeting-smartphones-with-css.html
Another rewrite.

The previous rewrite was done alongside fixing how my site handles mobile styles; both of those were done alongside hours of testing and searching to figure out the exact metrics involved; and I was focused on publishing the site fixes ASAP and rushed myself on writing and editing. The previous rewrite, then, is as to-the-point as the process that produced it: not at all. Hopefully this rewrite will be better.